FlaskDocker/requirements.txt
docker==4.4.1
Flask==1.1.2
Flask-WTF==0.14.3
WTForms==2.3.3
pyyaml==5.4
